Enable JAX float64 precision for numerical stability
Add jax_enable_x64 configuration at package init and upgrade remaining
float32 casts to float64 in Kalman update masks, Rao-Blackwell weights,
SVI losses, and missing-data preprocessing. The bulk of the float64
migration landed in the prior commit alongside the import rewrites.

+++ apps/data-pipeline/src/causal_ssm_agent/models/ssm/inference/__init__.py
@@ -787,7 +787,7 @@ def _fit_svi(
         losses.append(loss)
 
     svi_losses = (
-        jnp.asarray(losses, dtype=jnp.float32) if losses else jnp.empty((0,), dtype=jnp.float32)
+        jnp.asarray(losses, dtype=jnp.float64) if losses else jnp.empty((0,), dtype=jnp.float64)
     )
     svi_params = svi.get_params(svi_state)
